+++ scripts/common/watcher.js
@@ -5,5 +5,5 @@ const { flags } = require("./flags");
 
 exports.watcher = ts.createSolutionBuilderWithWatch(host, ["tsconfig.json"], {
   force: flags.force,
-  verbose: flags.verbose
+  verbose: flags.verbose,
 });